Commit the Final Results
Lesson 3.6.4
All of the coding seemed to be working well. Then, I misunderstood the directions in Lesson 5 and removed a bunch of code. After that, my file commands were not working properly and would not open, create and write data. I manually added the files and that seemed to help. However, the last part of this code is not working and I'm not sure how to fix it. Also, I have been saving my work in this repo all along so when I tried to go thru GitBash it would not accept the files.